+++ .idea/workspace.xml
@@ -2,7 +2,21 @@
 <project version="4">
   <component name="ChangeListManager">
     <list default="true" id="c095bca7-bdc0-4f79-aa31-85ca659691d6" name="Changes" comment="Update CD actions to fix PyAutoGUI issue">
+      <change beforePath="$PROJECT_DIR$/.idea/workspace.xml" beforeDir="false" afterPath="$PROJECT_DIR$/.idea/workspace.xml" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/main.py" beforeDir="false" afterPath="$PROJECT_DIR$/main.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/app.py" beforeDir="false" afterPath="$PROJECT_DIR$/src/app.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/settings.py" beforeDir="false" afterPath="$PROJECT_DIR$/src/settings.py" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/src/templates/forms.py" beforeDir="false" afterPath="$PROJECT_DIR$/src/templates/forms.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/templates/kv/commons.kv" beforeDir="false" afterPath="$PROJECT_DIR$/src/templates/kv/commons.kv" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/templates/kv/forms.kv" beforeDir="false" afterPath="$PROJECT_DIR$/src/templates/kv/forms.kv" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/templates/kv/progress_elements.kv" beforeDir="false" afterPath="$PROJECT_DIR$/src/templates/kv/progress_elements.kv" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/templates/kv/views.kv" beforeDir="false" afterPath="$PROJECT_DIR$/src/templates/kv/views.kv" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/templates/views.py" beforeDir="false" afterPath="$PROJECT_DIR$/src/templates/views/dashboard.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/tests/supports/unittest.py" beforeDir="false" afterPath="$PROJECT_DIR$/src/tests/supports/unittest.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/tests/test_index.py" beforeDir="false" afterPath="$PROJECT_DIR$/src/tests/test_index.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/tests/test_logon.py" beforeDir="false" afterPath="$PROJECT_DIR$/src/tests/test_logon.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/utils/directives.py" beforeDir="false" afterPath="$PROJECT_DIR$/src/utils/directives.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/utils/logger.py" beforeDir="false" afterPath="$PROJECT_DIR$/src/utils/logger.py" afterDir="false" />
     </list>
     <option name="SHOW_DIALOG" value="false" />
     <option name="HIGHLIGHT_CONFLICTS" value="true" />
@@ -45,22 +59,24 @@
   &quot;keyToString&quot;: {
     &quot;RunOnceActivity.OpenProjectViewOnStart&quot;: &quot;true&quot;,
     &quot;RunOnceActivity.ShowReadmeOnStart&quot;: &quot;true&quot;,
-    &quot;last_opened_file_path&quot;: &quot;D:/GitHub/Projects/bluech-client/src/templates&quot;,
+    &quot;last_opened_file_path&quot;: &quot;D:/GitHub/Projects/bluech-client/src/templates/views&quot;,
     &quot;settings.editor.selected.configurable&quot;: &quot;preferences.sourceCode.EditorConfig&quot;
   }
 }</component>
   <component name="RecentsManager">
     <key name="CopyFile.RECENT_KEYS">
+      <recent name="D:\GitHub\Projects\bluech-client\src\templates\views" />
+      <recent name="D:\GitHub\Projects\bluech-client\assets\fonts" />
       <recent name="D:\GitHub\Projects\bluech-client\src\templates" />
       <recent name="D:\GitHub\Projects\bluech-client\src\tests\supports\mocks" />
       <recent name="D:\GitHub\Projects\bluech-client" />
     </key>
     <key name="MoveFile.RECENT_KEYS">
+      <recent name="D:\GitHub\Projects\bluech-client\src\templates\components" />
+      <recent name="D:\GitHub\Projects\bluech-client\APPDATA\bluech" />
+      <recent name="D:\GitHub\Projects\bluech-client" />
       <recent name="D:\GitHub\Projects\bluech-client\docs\screenshots" />
       <recent name="D:\GitHub\Projects\bluech-client\src\templates\behaviors" />
-      <recent name="D:\GitHub\Projects\bluech-client\src\templates" />
-      <recent name="D:\GitHub\Projects\bluech-client\src\tests\supports" />
-      <recent name="D:\GitHub\Projects\bluech-client\src\helpers" />
     </key>
   </component>
   <component name="SpellCheckerSettings" RuntimeDictionaries="0" Folders="0" CustomDictionaries="0" DefaultDictionary="application-level" UseSingleDictionary="true" transferred="true" />
