+++ src/templates/kv/progress_elements.kv
@@ -33,7 +33,7 @@
         Color:
             rgba: self.ring_image_color
         Rectangle:
-            source: static.image(self.ring_image)
+            source: static_img(self.ring_image)
             pos: self.x + self.ring_trim + self.trim_extra, self.y + self.ring_trim + self.trim_extra
             size: self.width - self.trim_extra * 2 - self.ring_trim * 2, self.height - self.trim_extra * 2 - self.ring_trim * 2
